requirements.txt
annotated-types==0.7.0
anyio==4.9.0
beautifulsoup4==4.13.4
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
distro==1.9.0
duckduckgo_search==8.0.2
groq==0.4.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
lxml==5.4.0
primp==0.15.0
pydantic==2.11.5
pydantic_core==2.33.2
python-dotenv==1.0.0
requests==2.31.0
sniffio==1.3.1
soupsieve==2.7
typing-inspection==0.4.1
typing_extensions==4.14.0
urllib3==2.4.0
wikipedia==1.4.0
flask==2.3.3
openai==1.3.0
streamlit==1.28.0
gradio==3.50.0
langchain==0.0.350
numpy==1.24.3
pandas==1.5.3
